LayoutTests/http/tests/site-isolation/inspector/target/target-expected.txt


== Running test suite: SiteIsolation.Target
-- Running test case: SiteIsolation.Target.FrameTargetLifecycleEvents
{"eventType":"target-manager-target-added","targetType":"frame"}
{"eventType":"target-manager-target-added","targetType":"frame"}
{"eventType":"target-manager-target-removed","targetType":"frame"}
{"eventType":"target-manager-target-removed","targetType":"frame"}
